refactor: extract attacker turn body; round loop tolerates 0..N attacker turns (zero golden churn)
